inheirt whar we can from base class
