+++ modules/vaos/app/services/vaos/appointment_requests_service.rb
@@ -1,23 +1,9 @@
 # frozen_string_literal: true
 
-require_relative '../vaos/concerns/headers'
-
 module VAOS
   class AppointmentRequestsService < VAOS::BaseService
-    include Common::Client::Monitoring
-    include SentryLogging
-    include VAOS::Headers
-
-    configuration VAOS::Configuration
-
-    attr_accessor :user
-
-    STATSD_KEY_PREFIX = 'api.vaos'
-
     def self.for_user(user)
-      rs = VAOS::AppointmentRequestsService.new
-      rs.user = user
-      rs
+      VAOS::AppointmentRequestsService.new(user)
     end
 
     def get_requests(start_date = nil, end_date = nil)
