+++ modules/vaos/app/services/vaos/cc_supported_sites_service.rb
@@ -1,23 +1,7 @@
 # frozen_string_literal: true
 
-require_relative '../vaos/concerns/headers'
-
 module VAOS
   class CCSupportedSitesService < VAOS::BaseService
-    include Common::Client::Monitoring
-    include SentryLogging
-    include VAOS::Headers
-
-    configuration VAOS::Configuration
-
-    STATSD_KEY_PREFIX = 'api.vaos'
-
-    attr_reader :user
-
-    def initialize(user)
-      @user = user
-    end
-
     def get_supported_sites(site_codes)
       with_monitoring do
         response = perform(:get, url(site_codes), nil, headers(user))
